ENH: support for new cameras and new views. Port from Slicer 3.4 branch. As mentioned on the mailing list, apologies for instabilities while these core changes are further tested. Fiducials do not seem to follow camera for example, this will be fixed. Send bugs to sebastien.barre@kitware.com; Thank you.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=10795
git-svn-id: http://svn.slicer.org/Slicer4/trunk@10795 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/CommandLineModule/vtkMRMLCommandLineModuleNode.cxx
@@ -419,7 +419,7 @@ int
 vtkMRMLCommandLineModuleNode
 ::GetNumberOfRegisteredModules ()
 { 
-  return RegisteredModules->size(); 
+  return (int)RegisteredModules->size(); 
 }
 
 const char*
